third_party/WebKit/LayoutTests/virtual/threaded/fast/canvas-toBlob/canvas-toBlob-invalid-expected.txt

Test the handling of invalid arguments in canvas toBlob().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canvas.toBlob(); threw exception TypeError: Failed to execute 'toBlob' on 'HTMLCanvasElement': 1 argument required, but only 0 present..
PASS canvas.toBlob(null); threw exception TypeError: Failed to execute 'toBlob' on 'HTMLCanvasElement': The callback provided as parameter 1 is not a function..
PASS canvas.toBlob(undefined); threw exception TypeError: Failed to execute 'toBlob' on 'HTMLCanvasElement': The callback provided as parameter 1 is not a function..
PASS canvas.toBlob(function() { finishOneAsyncCall(); }); did not throw exception.
PASS canvas.toBlob(function(blob) { finishOneAsyncCall(); }, 'image/jpeg', 500) did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
